ebdjango/requirements.txt
Django==2.1.1
pytz==2018.6
djangorestframework==3.9.0
tensorflow==1.11.0
Pillow==5.3.0
dlib==19.16.0
opencv-python==3.4.3.18
matplotlib==3.0.1
imutils==0.5.1
boto3==1.9.42
